add: add separate components
