+++ README.md
@@ -1,1 +1,10 @@
 # React.js-Projects
+
+
+In this project we learnt about React Essentials and 
+
+    1=> Learnt about Components
+    2=> Learnt about JSX
+    3=> Learnt about Props
+    4=> Learnt about useState
+    5=> Learnt by creating the React essential ui
